backend/requirements.txt
fastapi
uvicorn[standard]
pandas
python-multipart
sqlalchemy
pydantic
python-dotenv